WebStorage/tests/submissions/Microsoft/storage.htm

<!DOCTYPE HTML>
<html>
<head>
  <title>Web Storage Tests</title>
  <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
  <link rel="author" title="Microsoft" href="http://www.microsoft.com" />
  <link rel="help" href="http://dev.w3.org/html5/webstorage/#storage-0" />
</head>
<script src="http://www.w3c-test.org/resources/testharness.js"></script>
<script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>

<script>
   function webStorageTest0() {
       var szKey = "";
       var storage = window.localStorage;
       storage.clear();

       storage.setItem(szKey, "bar");
       assert_equals(storage.getItem(szKey), "bar");
   }
   test(webStorageTest0, "localStorage empty string is a valid key");

   function webStorageTest0() {
       var szKey = "";
       var storage = window.sessionStorage;
       storage.clear();

       storage.setItem(szKey, "bar");
       assert_equals(storage.getItem(szKey), "bar");
    }
    test(webStorageTest0, "sessionStorage empty string is a valid key");
</script>
</body>
</html>
